requirements.txt
Django==1.11.7
djangorestframework==3.9.0
drf-yasg==1.11.0
django-environ==0.4.1
django-extensions==1.7.6
django-model-utils==2.6.1
pytz==2016.10
django-filter==2.0.0
coreapi==2.3.3
coreschema==0.0.4
